Add files via upload
Suggested text:
How to Become a Wikipedian in 5 Steps
Pedagogical Principles
Wikipedibus is a structured, self-directed activity based on the game device "side quest" in a broader educational program. The goal is for the learner to gradually discover how to contribute to Wikipedia independently. The intervention of the teacher is not required, though one should be available, only in case of need.
Metadata